+++ CODEOWNERS
@@ -1415,8 +1415,8 @@ build.json @home-assistant/supervisor
 /tests/components/skybell/ @tkdrob
 /homeassistant/components/slack/ @tkdrob @fletcherau
 /tests/components/slack/ @tkdrob @fletcherau
-/homeassistant/components/sleep_as_android/ @tr4nt0r
-/tests/components/sleep_as_android/ @tr4nt0r
+/homeassistant/components/sleep/ @tr4nt0r
+/tests/components/sleep/ @tr4nt0r
 /homeassistant/components/sleepiq/ @mfugate1 @kbickar
 /tests/components/sleepiq/ @mfugate1 @kbickar
 /homeassistant/components/slide/ @ualex73
