+++ homeassistant/components/sleep/const.py
@@ -1,10 +1,10 @@
 """Constants for the Sleep as Android integration."""
 
-DOMAIN = "sleep_as_android"
+DOMAIN = "sleep"
 
 ATTR_EVENT = "event"
 ATTR_VALUE1 = "value1"
 ATTR_VALUE2 = "value2"
 ATTR_VALUE3 = "value3"
 
-SLEEP_AS_ANDROID_EVENT = "sleep_as_android_event"
+SLEEP_EVENT = "sleep_event"
