+++ homeassistant/components/sleep/event.py
@@ -15,7 +15,7 @@ from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
 from . import SleepAsAndroidConfigEntry
-from .const import ATTR_EVENT, SLEEP_AS_ANDROID_EVENT
+from .const import ATTR_EVENT, SLEEP_EVENT
 from .entity import SleepAsAndroidEntity
 
 PARALLEL_UPDATES = 0
@@ -156,7 +156,5 @@ class SleepAsAndroidEventEntity(SleepAsAndroidEntity, EventEntity):
         """Register event callback."""
 
         self.async_on_remove(
-            async_dispatcher_connect(
-                self.hass, SLEEP_AS_ANDROID_EVENT, self._async_handle_event
-            )
+            async_dispatcher_connect(self.hass, SLEEP_EVENT, self._async_handle_event)
         )
