+++ homeassistant/components/sleep/manifest.json
@@ -1,10 +1,10 @@
 {
-  "domain": "sleep_as_android",
+  "domain": "sleep",
   "name": "Sleep as Android",
   "codeowners": ["@tr4nt0r"],
   "config_flow": true,
   "dependencies": ["webhook"],
-  "documentation": "https://www.home-assistant.io/integrations/sleep_as_android",
+  "documentation": "https://www.home-assistant.io/integrations/sleep",
   "iot_class": "local_push",
   "quality_scale": "silver"
 }
