+++ mypy.ini
@@ -4416,7 +4416,7 @@ disallow_untyped_defs = true
 warn_return_any = true
 warn_unreachable = true
 
-[mypy-homeassistant.components.sleep_as_android.*]
+[mypy-homeassistant.components.sleep.*]
 check_untyped_defs = true
 disallow_incomplete_defs = true
 disallow_subclassing_any = true
