+++ tests/components/sleep/snapshots/test_event.ambr
@@ -39,7 +39,7 @@
     'original_device_class': None,
     'original_icon': None,
     'original_name': 'Alarm clock',
-    'platform': 'sleep_as_android',
+    'platform': 'sleep',
     'previous_unique_id': None,
     'supported_features': 0,
     'translation_key': <SleepAsAndroidEvent.ALARM_CLOCK: 'alarm_clock'>,
@@ -104,7 +104,7 @@
     'original_device_class': None,
     'original_icon': None,
     'original_name': 'Anti-snoring',
-    'platform': 'sleep_as_android',
+    'platform': 'sleep',
     'previous_unique_id': None,
     'supported_features': 0,
     'translation_key': <SleepAsAndroidEvent.ANTI_SNORING: 'antisnoring'>,
@@ -161,7 +161,7 @@
     'original_device_class': None,
     'original_icon': None,
     'original_name': 'Lullaby',
-    'platform': 'sleep_as_android',
+    'platform': 'sleep',
     'previous_unique_id': None,
     'supported_features': 0,
     'translation_key': <SleepAsAndroidEvent.LULLABY: 'lullaby'>,
@@ -218,7 +218,7 @@
     'original_device_class': None,
     'original_icon': None,
     'original_name': 'Sleep apnea',
-    'platform': 'sleep_as_android',
+    'platform': 'sleep',
     'previous_unique_id': None,
     'supported_features': 0,
     'translation_key': <SleepAsAndroidEvent.APNEA_ALARM: 'apnea_alarm'>,
@@ -277,7 +277,7 @@
     'original_device_class': None,
     'original_icon': None,
     'original_name': 'Sleep phase',
-    'platform': 'sleep_as_android',
+    'platform': 'sleep',
     'previous_unique_id': None,
     'supported_features': 0,
     'translation_key': <SleepAsAndroidEvent.SLEEP_PHASE: 'sleep_phase'>,
@@ -339,7 +339,7 @@
     'original_device_class': <EventDeviceClass.BUTTON: 'button'>,
     'original_icon': None,
     'original_name': 'Sleep tracking',
-    'platform': 'sleep_as_android',
+    'platform': 'sleep',
     'previous_unique_id': None,
     'supported_features': 0,
     'translation_key': <SleepAsAndroidEvent.SLEEP_TRACKING: 'sleep_tracking'>,
@@ -402,7 +402,7 @@
     'original_device_class': None,
     'original_icon': None,
     'original_name': 'Sound recognition',
-    'platform': 'sleep_as_android',
+    'platform': 'sleep',
     'previous_unique_id': None,
     'supported_features': 0,
     'translation_key': <SleepAsAndroidEvent.SOUND_EVENT: 'sound_event'>,
